Moved 2 tests to ts
